[#286] Don't allow server-side initialization of any pages
