+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "windrift",
-    "version": "2.1.0",
+    "version": "2.1.1",
     "private": true,
     "scripts": {
         "build": "next build",
